+++ Project.toml
@@ -14,12 +14,15 @@ TypedTables = "9d95f2ec-7b3d-5a63-8d20-e2491e220bb9"
 
 [weakdeps]
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
+Chairmarks = "0ca39b1e-fe0b-4e98-acfc-b1656634c4de"
 
 [extensions]
 BenchmarkToolsExt = "BenchmarkTools"
+ChairmarksExt = "Chairmarks"
 
 [compat]
 BenchmarkTools = "1.5.0"
+Chairmarks = "1.2"
 CoverageTools = "1"
 CpuId = "0.3.1"
 TypedTables = "1"
