+++ ext/ChairmarksExt/ChairmarksExt.jl
@@ -1,11 +1,11 @@
-module ChainmarksExt
+module ChairmarksExt
 
 using Chainmarks
 using PerfChecker
 import TypedTables: Table
 
 include("b.jl")
-include("be.jl")
-include("chairmark.jl")
+#include("be.jl")
+#include("chairmark.jl")
 end
